add footer component and move signup comp to folder
