+++ client/src/App.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import Footer from './components/Footer';
 import Home from './components/Home';
 import NavigationBar from './components/NavigationBar';
 
@@ -9,6 +10,7 @@ function App() {
    <>
     <NavigationBar />
     <Home/>
+    <Footer/>
    </>
   </div>
  );
