+++ client/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import App from './App';
-import SignUp from './SignUp';
+import Signup from './components/SignUp';
 import './App.scss'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
@@ -11,7 +11,7 @@ root.render(
   <BrowserRouter>
    <Routes>
     <Route path="/" element={<App />} />
-    <Route path="/signup" element={<SignUp />} />
+    <Route path="/signup" element={<Signup />} />
    </Routes>
   </BrowserRouter>
  </React.StrictMode>
